sr: use linkResolver in getRoutesToGenerate
